% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRC_PARMS_02.R
\name{LRC_PARMS_02}
\alias{LRC_PARMS_02}
\title{Fit a Light Response Curve (LRC) by an index to get a parameter file}
\usage{
LRC_PARMS_02(
  data.frame = NULL,
  iterations = NULL,
  priors.lrc = brms::prior(normal(-0.01, 0.1), nlpar = "a1", lb = -0.2, ub = 0) +
    brms::prior(normal(-7.65, 0.33), nlpar = "ax", lb = -30, ub = -5) +
    brms::prior(normal(2.1, 0.11), nlpar = "r", lb = 1.9, ub = 2.2) +
    brms::prior(normal(25, 25), nlpar = "theta", lb = 0, ub = 3),
  idx.colname = NULL,
  NEE.colname = NULL,
  PAR.colname = NULL
)
}
\arguments{
\item{data.frame}{(dataframe) A dataframe that contains net ecosystem exchange (NEE), an index, and photosynthetically active radiation (PAR).}

\item{iterations}{(numeric) The number of iterations to run \code{brms::brm()}.}

\item{priors.lrc}{(brmsprior dataframe) The priors for \code{brms::brm()} to use.
Default priors are as follows:

\if{html}{\out{<div class="sourceCode">}}\preformatted{brms::prior(normal(-0.01, 0.1), nlpar = "a1", lb = -0.2, ub = 0) +
brms::prior(normal(-7.65, 0.33), nlpar = "ax", lb = -30, ub = -5) +
brms::prior(normal(2.10, 0.11), nlpar = "r", lb = 1.9, ub = 2.2)+
brms::prior(normal(25, 25), nlpar = "theta", lb = 0, ub = 3)
}\if{html}{\out{</div>}}}

\item{idx.colname}{(character) The name of the column containing the index.}

\item{NEE.colname}{(character) The name of the column containing NEE.}

\item{PAR.colname}{(character) The name of the column containing PAR.}
}
\value{
(dataframe) Dataframe of parameter values by the index used to fit them.
}
\description{
This function uses the equation:
\deqn{\text{NEE} \sim \frac{a_1 \cdot \text{PAR} + a_x - \sqrt{(a_1 \cdot \text{PAR} + a_x)^2 - 4 \cdot \text{PAR} \cdot a_1 \cdot a_x}}{2 \cdot \Theta} - r}

Where \eqn{r} is ecosystem respiration (𝜇mol CO2 m-2 s-1),
\eqn{a_1} is the apparent quantum efficiency of CO2 uptake (CO2),
\eqn{\Theta} is the convexity (curvilinear angle) of the nonrectangular hyperbola (degrees)
and \eqn{a_x} is the maximum CO2 uptake rate on the ecosystem scale.

The equation requires photosynthetically active radiation, PAR, in 𝜇mol m-2 s-1 and net ecosystem exchange, NEE, in 𝜇mol m-2 s-1.
}
\details{
Model parameters Are fit using the R package \code{brms}.

Rhat (Potential Scale Reduction Factor):
Indicates how well the different Markov chains in your analysis have converged to the same posterior distribution.
Ideally, Rhat should be close to 1 for all parameters.
A high Rhat value suggests potential convergence issues and the need to run the chains longer.

Bulk ESS (Effective Sample Size - Bulk):
Estimates the effective number of independent samples from the central part of the posterior distribution.

Tail ESS (Effective Sample Size - Tail):
Estimates the effective number of independent samples from the tails of the posterior distribution.
Important for assessing the reliability of quantile estimates (e.g., 95\% confidence intervals).

Key points to remember:
Aim for Rhat close to 1 and high values for both Bulk ESS and Tail ESS.
}
\examples{
# Write examples here

}
