% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRC_PARMS_03.R
\name{LRC_PARMS_03}
\alias{LRC_PARMS_03}
\title{Fit a Light Response Curve (LRC) by an index to get a parameter file}
\usage{
LRC_PARMS_03(
  data.frame = NULL,
  iterations = NULL,
  priors.lrc = brms::prior("normal(-9.9, 0.4)", nlpar = "ax", lb = -15, ub = 1) +
    brms::prior("normal(-0.05, 0.1)", nlpar = "a1", lb = -0.2, ub = 0.5) +
    brms::prior("normal(1.2, 0.11)", nlpar = "r", lb = 1, ub = 3.2),
  idx.colname = NULL,
  NEE.colname = NULL,
  PAR.colname = NULL
)
}
\arguments{
\item{data.frame}{(dataframe) A dataframe that contains net ecosystem exchange (NEE), an index, and photosynthetically active radiation (PAR).}

\item{iterations}{(numeric) The number of iterations to run \code{brms::brm()}.}

\item{priors.lrc}{(brmsprior dataframe) The priors for \code{brms::brm()} to use.
Default priors are as follows:

\if{html}{\out{<div class="sourceCode">}}\preformatted{brms::prior("normal(-9.9, 0.4)", nlpar = "ax", lb = -15, ub = 1) +
brms::prior("normal(-0.05, 0.1)", nlpar = "a1", lb = -0.2, ub = 0.5) +
brms::prior("normal(1.2, 0.11)", nlpar = "r", lb = 1.0, ub = 3.2)
}\if{html}{\out{</div>}}}

\item{idx.colname}{(character) The name of the column containing the index.}

\item{NEE.colname}{(character) The name of the column containing NEE.}

\item{PAR.colname}{(character) The name of the column containing PAR.}
}
\value{
(dataframe) Dataframe of parameter values by the index used to fit them.
}
\description{
This function uses the equation:
\deqn{\text{NEE} \sim a_x \left( 1 - \exp \left( -\frac{a_1 \cdot \text{PAR}}{a_x} \right) \right) - r}

Where NEE is net ecosystem exchange in \eqn{\mu}mol m-2 s-1,
\eqn{a_x} is light-saturated net photosynthetic rate,
\eqn{a_1} is the initial quantum efficiency or intrinsic quantum yield,
PAR is Photosynthetically active radiation in \eqn{\mu}mol m-2 s-1,
and \eqn{r} is dark respiration rate.

The equation requires photosynthetically active radiation, PAR, in \eqn{\mu}mol m-2 s-1 and net ecosystem exchange, NEE, in \eqn{\mu}mol m-2 s-1.

Reference: https://link.springer.com/article/10.1007/s12010-015-1543-0.
Equation # 8 in the paper (reference).
}
\details{
Model parameters are fit using the R package \code{brms}.

Rhat (Potential Scale Reduction Factor):
Indicates how well the different Markov chains in your analysis have converged to the same posterior distribution.
Ideally, Rhat should be close to 1 for all parameters.
A high Rhat value suggests potential convergence issues and the need to run the chains longer.

Bulk ESS (Effective Sample Size - Bulk):
Estimates the effective number of independent samples from the central part of the posterior distribution.

Tail ESS (Effective Sample Size - Tail):
Estimates the effective number of independent samples from the tails of the posterior distribution.
Important for assessing the reliability of quantile estimates (e.g., 95\% confidence intervals).

Key points to remember:
Aim for Rhat close to 1 and high values for both Bulk ESS and Tail ESS.
}
\examples{
# Import flux tower data
tower.data <- read.csv(system.file("extdata", "AMF_US-Skr_BASE_HH_2-5_Formatted.csv",
                                   package = "CarbonExchangeParameters"))

# Fit curve parameters for each YearMon:
Example_LRC_PARMS_03 <- LRC_PARMS_03(data.frame = tower.data,
                                     iterations = 1000,
                                     priors.lrc = brms::prior("normal(-9.9, 0.4)",
                                                    nlpar = "ax", lb = -15, ub = 1) +
                                                  brms::prior("normal(-0.05, 0.1)",
                                                    nlpar = "a1", lb = -0.2, ub = 0.5) +
                                                  brms::prior("normal(1.2, 0.11)",
                                                    nlpar = "r", lb = 1.0, ub = 3.2),
                                     idx.colname = 'YearMon',
                                     NEE.colname = 'NEE_PI',
                                     PAR.colname = 'SW_IN')


}
