% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRC_PARMS_05.R
\name{LRC_PARMS_05}
\alias{LRC_PARMS_05}
\title{Fit a Light Response Curve (LRC) by an index to get a parameter file}
\usage{
LRC_PARMS_05(
  data.frame = NULL,
  iterations = NULL,
  priors.lrc = brms::prior(normal(-0.12, 0.1), nlpar = "a", lb = -0.2, ub = 0) +
    brms::prior(normal(0, 1), nlpar = "B", lb = -1, ub = 1) + brms::prior(normal(0.2,
    0.1), nlpar = "Y", lb = 0, ub = 0.3) + brms::prior(normal(0, 0.1), nlpar = "Z", lb =
    -0.1, ub = 0.2),
  idx = NULL,
  nee = NULL,
  PAR = NULL
)
}
\arguments{
\item{data.frame}{(dataframe) A dataframe that contains net ecosystem exchange (NEE), an index, and photosynthetically active radiation (PAR).}

\item{iterations}{(numeric) The number of iterations to run \code{brms::brm()}.}

\item{priors.lrc}{(brmsprior dataframe) The priors for \code{brms::brm()} to use.
Default priors are as follows:

\if{html}{\out{<div class="sourceCode">}}\preformatted{brms::prior(normal(-0.12, 0.1), nlpar = "a", lb = -0.2, ub = 0) +
brms::prior(normal(0, 1), nlpar = "B", lb = -1, ub = 1) +
brms::prior(normal(0.2, 0.1), nlpar = "Y", lb = 0, ub = 0.3) +
brms::prior(normal(0.0, 0.1), nlpar = "Z", lb = -0.1, ub = 0.2)
}\if{html}{\out{</div>}}}

\item{idx}{(character) The name of the column containing the index.}

\item{nee}{(character) The name of the column containing NEE.}

\item{PAR}{(character) The name of the column containing PAR.}
}
\value{
(dataframe) Dataframe of parameter values by the index used to fit them.
}
\description{
This function uses the equation:
\deqn{\text{NEE} \sim -a_1 \cdot \exp(-B \cdot \text{PAR}) - Y \cdot \exp(Z \cdot \text{PAR})}

Where B and Y are correction factors in which
B is the photoinhibitation item,
Y is the light saturation item in which Y = a/Pmax,
and Z is a correction factor not well defined within the paper.

The equation requires photosynthetically active radiation, PAR, in 𝜇mol m-2 s-1 and net ecosystem exchange, NEE, in 𝜇mol m-2 s-1.
}
\details{
Model parameters are fit using the R package \code{brms}.

Rhat (Potential Scale Reduction Factor):
Indicates how well the different Markov chains in your analysis have converged to the same posterior distribution.
Ideally, Rhat should be close to 1 for all parameters.
A high Rhat value suggests potential convergence issues and the need to run the chains longer.

Bulk ESS (Effective Sample Size - Bulk):
Estimates the effective number of independent samples from the central part of the posterior distribution.

Tail ESS (Effective Sample Size - Tail):
Estimates the effective number of independent samples from the tails of the posterior distribution.
Important for assessing the reliability of quantile estimates (e.g., 95\% confidence intervals).

Key points to remember:
Aim for Rhat close to 1 and high values for both Bulk ESS and Tail ESS.
}
\examples{
# Write examples here

}
