% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TRC_PARMS_03.R
\name{TRC_PARMS_03}
\alias{TRC_PARMS_03}
\title{Fit a Temperature Response Curve (TRC) by an index to get a parameter file}
\usage{
TRC_PARMS_03(
  data.frame = NULL,
  iterations = NULL,
  priors.trc = brms::prior("normal(0.08, 0.03)", nlpar = "a", lb = 0, ub = 0.9) +
    brms::prior("normal(24, 4)", nlpar = "Topt", lb = 10, ub = 40) +
    brms::prior("normal(3.8, 5)", nlpar = "ERmax", lb = 0, ub = 10),
  idx.colname = NULL,
  NEE.colname = NULL,
  PAR.colname = NULL,
  TA.colname = NULL
)
}
\arguments{
\item{data.frame}{(dataframe) A dataframe that contains net ecosystem exchange (NEE), an index, air temperature (TA), and photosynthetically active radiation (PAR).}

\item{iterations}{(numeric) The number of iterations to run \code{brms::brm()}.}

\item{priors.trc}{(brmsprior dataframe) The priors for \code{brms::brm()} to use.
Default priors are as follows:

\if{html}{\out{<div class="sourceCode">}}\preformatted{brms::prior("normal(0.08, 0.03)", nlpar = "a", lb = 0, ub = 0.9) +
brms::prior("normal(24, 4)", nlpar = "Topt", lb = 10, ub = 40) +
brms::prior("normal(3.8, 5)", nlpar = "ERmax", lb = 0, ub = 10)
}\if{html}{\out{</div>}}}

\item{idx.colname}{(character) The name of the column containing the index.}

\item{NEE.colname}{(character) The name of the column containing NEE.}

\item{PAR.colname}{(character) The name of the column containing PAR.}

\item{TA.colname}{(character) The name of the column containing air temperature.}
}
\value{
(dataframe) Dataframe of parameter values by the index used to fit them.
}
\description{
This function uses the equation:
\deqn{\text{NEE} \sim a \left( T - T_{\text{opt}} \right)^2 + E_{\text{Rmax}}}

Where
\eqn{a} = a parameter estimate of the fitted quadratic function,
\eqn{T} = temperature in Celsius,
\eqn{T_{opt}} = optimum temperature, a parameter, vertex of the parabola,
and \eqn{E_{Rmax}} = maximum respiration, a parameter, vertex of the parabola.

Reference: https://www.nature.com/articles/s41559-023-02121-w#Equ1

The equation requires air temperature (TA) in degrees Celsius, photosynthetically active radiation (PAR) in \eqn{\mu}mol m-2 s-1, and net ecosystem exchange (NEE) in \eqn{\mu}mol m-2 s-1.
}
\details{
Model parameters are fit using the R package \code{brms}.

Rhat (Potential Scale Reduction Factor):
Indicates how well the different Markov chains in your analysis have converged to the same posterior distribution.
Ideally, Rhat should be close to 1 for all parameters.
A high Rhat value suggests potential convergence issues and the need to run the chains longer.

Bulk ESS (Effective Sample Size - Bulk):
Estimates the effective number of independent samples from the central part of the posterior distribution.

Tail ESS (Effective Sample Size - Tail):
Estimates the effective number of independent samples from the tails of the posterior distribution.
Important for assessing the reliability of quantile estimates (e.g., 95\% confidence intervals).

Key points to remember:
Aim for Rhat close to 1 and high values for both Bulk ESS and Tail ESS.
}
\examples{
# Write examples here

}
